+++ test/runtests.jl
@@ -12,6 +12,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_clusters.jl")
         include("unit/test_salc.jl")
         include("unit/test_fit.jl")
+        include("unit/test_torque.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")